Remove feature category from loadbalancer
This removes the feature category information from loadbalancer
SLIs. We do this, because having a feature category on these would
also feed the SLI into the error budget for stage groups, and
eventually route alerts to the stage group's configured Slack
channel. But stage groups receiving these alerts aren't likely to know
how to resolve issues with this component.