fix: glsh update command
Background
---
When running `glsh update` command it was failing with the error below:
```
fatal: not a git repository (or any of the parent directories): .git
```

This is because we where running a git command for `CURRENT_BRANCH` but
we didn't change the directory to the runbook directory.

Solution
---
Before setting the variables, make sure we change the directory to the
runbook directory.