Add back global recordings for application SLIs
This will use the existing recordings in Prometheus if they are
present: which would happen if the SLI is already used in the service
catalog an is added to the `recordingRuleMetrics` stanza.

If this is not the case, we'll use the recording from the source
aggregation set.

Doing this will make sure we always have the global aggregation set
available for dashboards, regardless if they are used in the service
catalog.