feat(dashboards): limit refresh intervals
Background
---
Most dashboards have lots of panels and send a lot of requests to
Thanos. Grafana has a feature called
[refresh](https://grafana.com/docs/grafana/latest/dashboards/time-range-controls/)
which automatically refreshes the dashboard for the user. The default
right now 5s, so every 5 seconds it sends all those queries to Thanos
which puts an unnecessary load on Thanos, and we saw this causing
problems in
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/14459
where 1 dashboard can cause an alert because it's overwhelming Thanos.

Solution
---
Limit the refresh intervals from 1m to 30m so we prevent 1 user DoSing
Thanos. If the user needs to have constant refreshes, they can
just click the refresh button when they need to.

Right now no-one will able to override this, and all dashboards will use
the same refresh internvals, if we see the need in the future we can
have this as a parameter which the calls for `basic.dashboard` can pass.

reference https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/14459