Add success rate recordings to aggregation sets
This adds the possibility to add a `successRate` field to aggregation
sets.

If this is present in a set for a target aggregation, we'll try to get
it directly from the source. If that does not exist, but an error and
ops-rate do, we'll calculate the `successRate` that way. If there is
no error rate but there is an ops rate, then we'll consider all
operations successful for the successRate.

This is how we're using the error rates in the stage group error
budgets. So this is more convenient to work with.

The upscaling also works the same way: using upscaled versions of the
ops- and error rates. But using a fallback of `0 * ops` if the error
rate is missing.