feat(grafana): disable auto refresh by default
Background
---
Most dashboards have lots of panels and send a lot of requests to
Thanos. Grafana has a feature called
[refresh](https://grafana.com/docs/grafana/latest/dashboards/time-range-controls/)
which automatically refreshes the dashboard for the user.

This puts a lot of strain on Thanos and it's services since we end up
refreshing the same panels for a long time when we don't really need
them too.

Solution
---
Disable the auto refresh by default, and let the user specfically opt-in
into automatically refreshing this.

Please note that this doesn't disable the auto-refresh on the un-managed
dashboards since those are hard coded in the database.

reference https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/14459