Account for missing error rate in error budget
Some SLIs only have an operation rate, in which case, the error rate
is missing. Because of the extra aggregation by component, these would not be
excluded anymore when displaying the error budget.

This accounts for that by replacing the missing error rate with a 0 if
the operation rate is present.

This would make sure that the traffic shares are still present.