Make serviceType nullable for dashboard panels
We are adding some panels where we want a series per type, rather than a
row per type. Those will mostly Just Work without a serviceType
parameter, as long as the selector is still sensible and no aggregation
is needed, but there were a few places where we assumed that parameter
must be present (mostly to add it to the selector hash).

This change prepares the ground for that future change.