Adds log links to the search service/dashboard
The application search elasticsearch itself logs to *another* hosted elasticsearch in the Elastic Cloud service.  And although this *service* does 'requests', the logs we can link to do not record those (they're operational logs for the cluster) and therefore there's no point including the 'requests' links, so this adds support for excluding those links in appropriate situations.