Add a dashboard opt-in status for groups
This adds a dashboards that shows what portion of traffic will be
included in the error budget for stage groups based on whether or not
they opted in.

It can be filtered to only show a selection of stages or groups.

The traffic by urgency metrics are queried using high cardinality
metrics. So this is limited to the last 6 hours, not the entire range
selected. The panels on the top, showing opt-in & opt-out status can
be queried over the range selected in Grafana.