Updates GCP quota saturation monitoring to use new label
In https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/tanka-deployments/-/merge_requests/242 and https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/tanka-deployments/-/merge_requests/243 we ensured that the original 'region' label from the gcp quota exporter was exposed all the way through to Thanos (not hidden by an 'external label' configuration that hard-coded 'region').

This closes the loop.  To be fair, for *current* deployments it's basically redundant because all resource usage for our monitored projects is in the us-east1 region, so the only metrics which will pass the division and conditions of the saturation metric will be ones in that region (if the limit is 0 we get NaN, and if the usage is 0 we fail the '> 0' condition).  The above MRs solve the immediate problem by expanding the metrics so the right metric *always* works.  But for best results in the event we gain resources in other regions in future, this MR ensures it *remains* correct and viable, no matter where we run the quota exporter.