Refactor: puma feature category metrics
Before this, we were recording feature category metrics for the puma
component based on the promSourceSLIs aggregation set. This means that
we'd have a 1m burn rate, which the feature categories don't need.

Because of this adding burnRates to the aggregation set rates would
not generate the rules for the puma component.

This refactors this to use the featureCategorySourceSLIs as the source
of truth for the generation of this aggregation. Similar to how we
generate the feature category recordings for Sidekiq.

This also means that we don't need to have these recordings inside the
key metric groups. Keeping that group smaller.

Because of that, we also have less rules to evaluate: instead of
recording per service and selecting based on the type label, we record
them and include the type as an aggregation label.