Include `env` in the join for feature_category
The `mapping` isn't always updated at the same time for all envs. For
example the `testbed` lags behind, meaning that when feature
categories or groups change, there are label mismatches when joining.

This solves that by including the env in the join. The testbed itself
won't have any feature category metrics populated.