Upscale burn rates in stage group aggregation
This allows the stage group aggregation set to use it's own 1h
recordings to upscale to the 6h & 3d burn rates.

This way, we do not have to check and aggregate all data from all of
the different Prometheus hosts that might or might not have this
metric.

This uses this approach for the feature category and stage group
aggregation sets first. We could expand these to other upscaled
sources too.