Generate stage group feature mapping in Thanos
This makes these metrics available with the Thanos ruler
`monitor='global'`. Having these available, will allow us to test if
adding that selector to the queries that can use it will speed them
up.

This means that we only need to query one store (the global one)
instead of aggregating all Prometheus stores that might be emitting
this recording.

This does not use this newly generated rule: we'd like to try it out
to compare performance first.