Use global feature category mapping
This uses the global feature category mapping to map feature
categories to stage groups in recordings and dashboards.

The advantage is that we will not be surprised by new labels being
added to Prometheus targets in Thanos, and some of the Prometheus
recording a mapping that is out of sync.

This could also impact performance of stage group recordings in
Thanos.