Split off long window feature category recordings
This splits the long window and short window feature category
recordings.

The short window recordings, specifically the 1h one, is what is used
most of the time for all and we need those every minute, because we
count on having 1 sample per minute.

The longer windows aggregate a lot more data, which make them slower.

For this aggregation set the longer windows here use upscaling of the
short window recordings. This means we can record at a lower interval,
making it less likely that the rule evaluation will take longer than
the interval

- The upscaling of rates is an `average_over_time`.
  For an average of an hourly rate over such a long window,
  the resulting values would not change significantly in a few
  minutes. So 5 minutes should be plenty.
- The upscaling of ratios uses a `sum_over_time`.
  For both the  numerator and the denominator we use recordings
  sourced from the same short window group. So we can be sure that
  we'll have the same number of samples on both sides of the fraction.

These recordings are currently unused.