Remove alert for stale cached GitLab-rails repo
We used to store an archive of the GitLab-rails repository in GCS to
allow fetching that archive and reducing the load on Gitaly.

We have since optimized the way Gitaly provides the objects to 100s of
jobs requesting the same thing.

We've stopped updating the archive in
https://gitlab.com/gitlab-org/gitlab/-/issues/344792

This alert is no longer necessary.