Allow excluding labels from SLI queries
When we create a dashboard per stage group, we want to aggregate SLIs
across services. This only gets you so far, and then you have to get out
and hop, because an SLI is tied to a service. In fact, our hierarchy is
Service > SLI > {error rate, request rate, apdex, etc.}.

In this case we treat a single SLI as representing all SLIs of the same
ID across all services. For instance, we can take the `puma` SLI from
`api`, `git`, or `web`, and then project it across the other two
services. This mostly works, but some queries add labels from the SLI
automatically in these dashboards, and so we need to support removing
them.

We do that with a new `withoutLabels` parameter that is pretty
ubiquitous - this change adds it to more places than strictly required -
that will exclude the given labels from the query we generate. That way
we can pass `withoutLabels=['type']` to these functions and get the
correct results back.