docs: update glsh install command to run sudo
Update the command to run with `sudo` rather then normal user since on
Linux `/usr/local/bin` only `root` can write to it.

The make command could be updated to run `sudo` automatically but it's
better the caller explicity gives it `sudo` permissions rather then
"hiding" it from the user.