Remove the apdex component from the puma SLI
This is now monitored using the `rails_requests` components, which
supports adjustable target durations set in the product.

This also moves the tooling links for slow requests into the
`rails_requests` component with a slow request duration of 5
seconds. That's currently the maximum target configurable in
GitLab-rails. We plan to make this endpoint specific.