Make error budget panels not rely on puma.apdex
Before we were using the `puma.apdex.satisfiedThreshold` value dug up
from the metrics catalog to build a Kibana vizualisation showing
requests not meeting the threshold.

This changes that to hardcode the 1s value next to the 5s value that
was already hardcoded. The values are still valuable as they are the
threshold for the `default` and `low` urgencies.