Exclude ops without error rate from error budget
This excludes operation rates from the error budget for stage groups
calculations for components that do not have an error rate.

The previous approach marked these operations as successes. Which is
not entirely correct, because we can't actually know if the operation
was a success.

Furthermore, all SLIs require an operation rate to be defined, but not
an apdex or an error rate. This means that the `rails_requests`
component we recently added would count all requests as a success on
that front.

This affects the following groups:

- global_search
- package
- pipeline_authoring
- pipeline_execution
- release
- workspace