docs: CustomersDot DB backups
Background
---
In
https://gitlab.com/gitlab-com/gl-infra/customersdot-ansible-poc/-/issues/56
we've added
https://ops.gitlab.net/gitlab-com/gl-infra/gitlab-restore/postgres-prdsub
to run database backups for `stgsub` and `prdsub`, which are the new GCP
environments for customers.gitlab.com.

Solution
---

Update the backups page to that into consideration the `prdsub`
environment since everything works the same.

This can feel like a shoehorn adding `prdsub` to `gprd` backup, but I
feel like the content is generic enough to reflect how we do backups in
all of postgres environments.