Record monthly availability by stage group
This adds recording rules for monthly availability by stage group.

This allows other clients (infradev-report, error-budget-reports,
periodic queries) to use these recordings rather than performing the
query themselves.

This means that we'll have a SSoT for the error budget calculation
which lives in this repository, and we'll have a historical record of
a groups performance with the calculation at the time.