Lower the deployment apdex SLO for the git service
This lowers the deployment SLO from 0.9995 to 0.999, which is still
tighter than the 0.998 SLO we use for alerting.

The reason for this is that the deployment SLO cannot take into
account customized SLOs specified on specific components in the
service. We've added the `rails_requests` service, and lowered the SLO
for that particular component while we work on improving the
signal (https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1343).

This came to light in a recent incident:
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6230 in
which the service would not meet the 0.9995 target and then block a
deploy or a feature flag roll out.

We should tighten this again after we've found the root cause of
slowdown in requests. But in the mean time, we should not block
deploys or feature flag roll outs because of it.