add an alerting rule for 404 errors
Deployments to docs.gitlab.com are scheduled every 4 hours, at 11:29 on
the 28th there was a deployment
https://gitlab.com/gitlab-org/gitlab-docs/-/pipelines/438110984/builds
incidents started happening at around the same time this went on for
approximately one hour. This MR will create an alerting rule that will
be triggered if there is a status of 404 on docs.gitlab.com for more
than 30 minutes.

Related incident:
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6125
Resolves corrective action:
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/14872