Correct maxLength for shortening dashboard uids
We were using an incorrect max length for truncating the stage group
dashboard identifier:

The identifier for the regular dashboard is generated from the
filename and directory the dashboard resides in. These filenames are
truncated when they are generated

The identifier for the shared detail dashboard comes from the
generated JSON. So at the moment we compile the dashboard.

This changes the `dashboardUid` function so that it would return the
same truncated uid that we'd use from the truncated filename.

The only group with a name that doesn't fit the uid at this time is
Authentication and Authorization.