doc: introduce gameday scenarios
Situation
---
We have incidents like
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/15104
because we have an assumption that something works, however in reality
it's not.

Complication
---
We end up with broken tools that we think are working but aren't which
can lead to slower recovery time because when we go and use the tool
we find out it's broken and have to spend time fixing it first.

Question
---
How can we make sure that recovery tools that rely on, work, and are
verified as expected, and have sufficient monitoring on verification
processes?

Answer
---
Thanks to @andrewn suggestion in starting to introduce [game
days](https://www.gremlin.com/community/tutorials/introduction-to-gamedays/)
where we stop/break something in purpose and see if works as expected.

Some inspiration taken from the following resources:

- https://www.gremlin.com/gameday/
- https://www.gremlin.com/community/tutorials/introduction-to-gamedays/
- https://www.youtube.com/watch?v=xs8OLmlMhVg

ref https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/15104