Separate recording rules by burn type in Thanos
This separates the recording rules in Thanos in 2 groups by burn type,
similar to what we do for source metrics in Prometheus.

This reduces the number of rule evaluations that Thanos-ruler needs to
perform in a single group. As a result, the recording of metrics used
on dashboards will not suffer from slow evaluation of large datasets
for long windows. Because the rule evaluation of bigger datasets now
happens over a longer interval, we also save some CPU cycles.

There are 2 burn types:

- Fast: recorded over a 1m interval
- Slow: recorded over a 2m interval.

The burn type for default burnRates is configured in
mwmbr/multiburnFactors.windows. By default only the 5m/1h burn rate is
considered a fast burn. Other windows, 30m/6h and 6h/3d, are considered
slow burns.

When recording for a different window, the interval will be the one
for a fast burn if the window is < 1h, otherwise it will be the
interval for a slow burn.