Add health status descriptions for stage groups
This adds a health status description to the stage group error budget
detail dashboard.

The health status displays the state based on multiple windows: when
the a single window of an alerting threshold is exceeded, we display
"Warning". When both the short and long windows are exceeded we
display "Degraded".

These panels for services use the overall service SLO, which is
recorded in Thanos. For stage groups, we use a fixed 99.95%
threshold.