Unify the rails and rails_api indexes in Kibana
Since the log messages for requests look the same for API requests and
web/git requests, this unifies the indexes. This will make these
toolinglinks work for dashboards that don't specify a type (for
example for stage groups).

For the service dashboards, the `type` label is part of the
selectorHash for the SLI, so we'll still filter for that on those
dashboards.