Use source metrics for recording error ratios
An aggregation set defining an error rate does not mean it is present
for all label combinations.

For an SLI definition it is mandatory to have an operation rate. But
not all SLIs have an error rate. So when calculating an error ratio,
we need to ignore these operation rates.

We can do this by always recording these ratios from the
source-aggregation set, and filtering out the values that don't have
an error rate on the bottom of the fraction.