Add a Kibana table aggregating apdex & errors
This table uses some painless scripting to show the sum of errors and
the slow requests and order by that. This means that stage groups can
look at this table and the top endpoint would be the most valuable to
work on.

This also needed to make sure that we correctly encode `+` in the
Rison.