Use region as significant label for webservices
This uses `region` instead of `fqdn` as one of the significant labels
in the `web`, `git` and `api` services. These services have been fully
migrated to Kubernetes. So they don't have an fqdn, but they do have a
region.

This will make it easier for us to see if an SLIs impact is regional
in the SLI details panels.

The changed SLIs here are:
- puma (`web`, `api`, `git`)
- imagescaler (`web`)
- workhorse (`web`, `api`, `git`)