Increase alert threshold for PostgreSQL_PendingWALFilesTooHigh
In the last few weeks, the Postgres write rate has increased to the
point where wal-g cannot always keep up with archiving the WAL files.

Efforts to analyze and improve this saturation problem are being
tracked here:
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15362

As of yesterday, we have increased the WAL archiving throughput
ceiling from 4 files/second to up to 6 files/second.  This
has helped by:
* Reduced the odds of accumulating a backlog.
* Reduced the time to resolve a backlog when one does accumulate.

That said, we are still seeing an elevated rate of WAL generation,
so we are still getting intermittent short-lived alerts.  The normal
pattern exhibits a series of short-lived backlogs that resolve quickly,
and the size of those backlogs now tends to spike slightly above the
alerting threshold (2K pending WAL files).

This MR increases that threshold to 3K, to reduce alert fatigue.
We keep the same 5-minute window, so that we will still get a prompt
alert if the backlog starts growing rapidly.