Use last sample for alerts and health checks
This makes us use the last sample for alerts and health checks.

This way, if an alert goes pending, it does not get canceled if we are
missing metrics. Or a health check will not ignore a service that was
temporarily not reporting metrics.

We could be mising metrics because Prometheus becomes unavailable,
or because recordings fail.