Use urgency from metrics for Sidekiq apdex
This makes sure we always use the `urgency` on metrics for determining
if a duration met the target, regardless of the shard that picked up
the job.

This also marks the `memory-bound`, `imports` and `quarantine` shard
as non-throttled ones. Don't have a throttled worker-attribute, so we
should be paying attention to the queue time and the saturation
there.