Remove successRate for stage group aggregations
This also removes the functionality to record a success rate from an
error rate. Since this was the only place this was used.

This is effectively a revert of
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/4089

The success-rate recorded from an error rate that is itself already a
recording means the success rate was "lagging". When summing this up
over a period of 28d, this accumulates and causes a bigger difference
then expected.

This was initially to not have to account for a missing error rate in
the numerator of the stage-group availability query. But we'd still
have to solve the problem by excluding a missing success-rate in the
denominator. So this wouldn't not solve a whole lot.