Lower the Google LB SLI for logging service
This lowers the SLI for the LB in front of log.gprd.gitlab.net. The
service doesn't receive a lot of traffic, just enough to not be
ignored. But this still means that a handful of failures push it over
the line for slow burns.

This lowers the error ratio SLO for that particular service from 99.9%
to 99.5% to make the alert less trigger happy.