Gitaly: Exclude RepositorySize from appdex
RepositorySize is changing the way it calculates size. It will now use
git-rev-list(1) instead of doing a `du -sk`. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4430

This RPC will not have a similar performance profile as RepackFull, as
it will do a full Git object database walk.

This RPC is also only called by background workers in Rails, so it makes
sense to not include its latency in the appdex calculation.

Changelog: changed