Runbook for recovering if CI Patroni cluster fails as standby replica
This was implemented as an action item from
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15683#note_952277309
where we determined there is a real risk that our CI Patroni replica may
fall sufficiently far behind that it cannot recover and this can lead to
a cascading situation where we send all `read-only` traffic to our
Primary which could lead to CPU saturation and an outage.

This documentation is based on the same plan we were already going to
use [for rollback in case something goes wrong with CI
decomposition](https://gitlab.com/gitlab-org/gitlab/-/issues/361759) but
it can also give us comfort in case something goes wrong before then due
to how our CI Patroni cluster is configured.