Exclude elastic_jvm_heap_memory from capacity planning
This resource should not saturate completely: the JVM manages the heap
size. If this saturates, then either:

1. Something has gone terribly wrong with the JVM.
2. We have misconfigured something.

Both of those will happen much quicker than a week, but we only run the
capacity planning report once a week. We will still keep the component
in the catalog, so saturation is visible on the regular dashboards.