Support combining rateApdex metrics
This correctly checks if the underlying metrics are backed by a
histogram. That way we can support generating `histogram_quantile`
graphs for those. While skipping the generation of that graph for
combined `rateApdex` SLIs.

We were already skipping that before, because `rateApdex` does not
have a `percentileLatencyQuery` field.

For the existing combined histogram apdexes, it makes sure we
correctly concatenate the aggregation labels.