Allow specifying a fallback in application SLIs
This allows specifying a `fallback` SLI for an application SLI. Doing
so allows us to specify different metric names to fall back to in
recording rules. Doing so, will `combine` both counters in the SLI,
the fallback being second in line.

This also makes the separator configurable, and default to `_`. The
correct implementation.

Using this, we've also created a default fallback SLI that uses `:` as
the separator.

This means that we'll look for metrics with a `_` first, and fall back
to the ones with an `:`. This will allow us to transition metric names
without losing the observability during a deploy.