Don't always double aggregate combined histograms
We need to do this for histogram apdexes that define both a satisfied
and a tolerated threshold. The sum of 2 metrics with a tolerated & a
satisfied thresholds will fail because the uniqueness labels don't
match. So instead we sum before we apply the uniqueness labels.

This means we don't match the recording rules anymore for those
metrics. So we only do this when needed.