Only use specific labels in capacity planning alerts
Some of the other labels on these series are prone to change in a way
that would generate a new alert, but would not actually represent a new
event.

For example, the `instance` label is Tamland-internal and used for
managing pushes to the push gateway. When we change this (by moving
items from one page to another, or re-writing how we handle groups of
metrics internally), we would get a new alert, but the
underlying (component, type) pair would not have changed.