Always expand all services a saturation point applies to
`open_fds` was the only saturation point using `allExcept`, which
expanded to a negated selector in Prometheus, and a nested object in
JSON. This doesn't work well for Tamland, which assumes the list of
services will always be fully expanded.

As this was only used in one place, just remove the option to use
`allExcept` entirely. Instead, `metricsCatalog.findServicesExcluding`
can be used to much the same effect.