chore: Use metrics separated with a _ instead of a :
This is second first step in switching to metrics separated with an _
as recommended in the Prometheus best practices:

> Note: The colons are reserved for user defined recording rules. They
> should not be used by exporters or direct instrumentation.

https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels

This starts using the metrics that are currently being emitted from
the applications, since:

- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/90977
- https://gitlab.com/gitlab-org/customers-gitlab-com/-/merge_requests/4983

Currently the application is emitting both versions: separated with
`_` and separated with `:`. This starts using the metrics with a
`_`. After we merge this, we can remove the old version from the
application.