feat: show promql_template_1 link in alert
What
---
Add a link to the `promql_template_1` annotation if present to the alert
template.

Why
---
Before we refactored the template in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15765 and
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15894 we
used to show the `promql_template_1` attached to the alert which was a
more clear version of the metric for debugging what the problem is.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16015
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>