feat: combine gprd main and cny pages
What
---
- Combine the pagerduty pages for `stage` when it's `main` and `cny`
- Delete the now unused `RouteCase` method.

Alertmanager config diff:
```
$ git switch master
$ ./alertmanager/generate.sh && cp alertmanager/alertmanager.yml /tmp/alertmanager.old
$ git switch feat/combine-cny-main-alerts
$ ./alertmanager/generate.sh && cp alertmanager/alertmanager.yml /tmp/alertmanager.new
$ diff /tmp/alertmanager.new /tmp/alertmanager.old
1378c1422,1423
<   - group_by:
---
>   - continue: true
>     group_by:
1386a1432,1444
>     routes:
>     - continue: false
>       matchers:
>       - env="gprd"
>       - slo_alert="yes"
>       - stage="cny"
>       receiver: slo_gprd_cny
>     - continue: false
>       matchers:
>       - env="gprd"
>       - slo_alert="yes"
>       - stage="main"
>       receiver: slo_gprd_main
```

Why
---
In https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15765
we've started grouping alerts by the service this resulsted into group
the alerts with that are firing for `cny` and `main` stages which is
what we want. However we end up paging the on-call twice because we have
2 recivers as show in https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15964#problem.

We already show the `stage` if it's `main` or `cny` in the alert labels
and in the alert description so it's already pretty clear wich stage is
effected.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15964
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>