gitaly: Include upload-pack in SLOs again
With commit 2d11c493 (Gitaly error SLO: Temporary exclusion of
upload-pack due to high volume of "fatal: the remote end hung up
unexpectedly" classified as "Internal", 2022-07-19) we have started to
ignore errors returned by the git-upload-pack(1) RPCs. The reason is
that we have a constant rate of errors due to the user aborting the
fetch, which would result in Internal errors being returned by Gitaly.
Given that this is an expected outcome though and that we cannot do
anything about the user aborting a fetch it didn't make sense to count
this into our SLOs.

In the Gitaly project we have since grown the logic to instead label any
such errors as `Canceled` with an error message indicating that the user
has canceled the fetch. And because `Canceled` errors don't count into
the SLO we don't have to ignore these RPCs anymore.

Revert the temporary exclusion of git-upload-pack(1) RPCs.