feat: decrease pgbouncer client con SLO to 85%
What
---
Bring back the original 85% hard SLO, and 80% soft SLO

Why
---
We where reaching the `max_client_conn` saturation SLO form time to time
an until we addressed this in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16121#note_1051842221
we increased the SLO to stop paging the on-call.

As we can see in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16121#note_1051842221
the saturation was fix so we can decrease the SLO onces more.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16121
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>