feat: remove gitalyruby component
Only a handful of methods are still in gitaly ruby, making it a very
low traffic and noisy SLI.

The goserver SLI still includes these methods as well. This is called
before gitaly-go delegates to gitaly-ruby. This means we will not be
missing any grpc methods in our goserver SLI.