feat: merge goserver_op_service into goserver
We used to have this SLI separate to be able to set lower targets for
all RPCs in that service. But because the SLI is quite low traffic,
it's also quite noisy.

Because of the changes we made to the `combined` helper. We can now
merge this SLI back into the goserver one.

This makes sure we no longer ignore errors for `OperationService` in
the main SLI. It also adds back `OperationService` apdex as a slow
one. Since now there will be more traffic in the combined SLI, this
should be less trigger happy.