Add redis-cache back to capacity planning
With https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/764, we've
gone from deliberately running redis-cache at maxmemory, to
understanding we don't want that (the evictions can cause latency
spikes) and therefore being overprovisioned.

This level of headroom - we're aiming to be below 70% - means that the
capacity planning reports on redis-cache will be useful again, so let's
add it back.