feat: active gitaly node disk space saturation
This adds a saturation point to monitor the amount of disk space that
is available for receiving new repositories

We're using the sum of all the disks of nodes with more than 20%
available disk space as a proxy for the number of active nodes. We
want the total available space of nodes with more than 20% available
to be at least 10% of the total space.

This would mean that we want approximately 10% of gitaly nodes to
receive new traffic.