Prepare Redis dashboards for Redis on Kubernetes
In Kubernetes deploys, we won't have a `fqdn` label. The `instance`
label is either `$fqdn:$port` (for VMs) or `$ip:$port` (for k8s), so
we need this for this case. However, the `$ip:$port` isn't very
friendly. So for Kubernetes, we also include the pod labels in
selectors and legends.