chore: Increase SLO for ruby_thread_contention
This increases the SLO for ruby thread contention from 75% to 85%.

We generally want this to be quite high, it means that we're fully
utilizing the single ruby-thread that can be active at the same
time. We just don't want it to be 100% because that might mean other
threads are starved.

In
https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/276
and
https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/274
we've seen that 85% is about the maximum we hover around. So this
would make those non-actionable issues less frequent.