fix: Feature category metrics generation from Sidekiq SLIs
This makes sure that we can generate feature category metrics from our
Sidekiq SLIs.

Before this change we would be generating the
`rules/autogenerated-feature-category-metrics-sidekiq.yml` file after
we had already generated the regular file. So the latter would take
presedence in the JSON object we geed into jsonnet-tool.

This moves those recordings to a new file, which means that the
recordings in the
`rules/autogenerated-feature-category-metrics-sidekiq.yml` are the
ones generated from the Sidekiq-SLIs directly. While the ones in
`rules/autogenerated-custom-feature-category-metrics-sidekiq.yml` are
the ones for Sidekiq jobs that we'll improve upon in
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/700

This was needed so stage groups can add their own SLIs to Sidekiq, for
example in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/5017